backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
anthropic>=0.7.0
akshare>=1.16.0
baostock>=0.8.8
tushare>=1.2.89
yfinance>=0.2.40
pandas>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.10.0
sqlalchemy>=2.0.0
